auth: forward authentication from hash to route handlers

+++ services/api/src/index.ts
@@ -18,7 +18,12 @@ import {
   reindexCategory,
   updateCategoryLeaveWithSlug,
 } from "./model/category"
-import { buildHandlers, getPaginatedParams } from "./utils/buildHandlers"
+import {
+  buildHandlers,
+  getPaginatedParams,
+  make401,
+  make403,
+} from "./utils/buildHandlers"
 import { deletePictureById, getPictureById } from "./model/picture"
 import { toCategoryApi, toCategoryApis } from "./controllers/categories"
 import { toPictureApi, toPictureApis } from "./controllers/picture"
@@ -29,18 +34,22 @@ import {
   deleteCategoryLink,
   deletePicturesCategoryLink,
 } from "./model/category/parents"
+import { AuthRole, validateOrNull } from "@local/auth"
 
-const runHandlers = buildHandlers({
-  CATEGORY: async ({ args: { slug } }) => {
+const runHandlers = buildHandlers<RouterContext>({
+  CATEGORY: async ({ args: { slug }, context }) => {
+    if (!context.role) return make401()
     return await toCategoryApi(await getCategoryLeaveWithSlug(db, slug))
   },
-  CATEGORY_UPDATE: async ({ args: { slug }, json }) => {
+  CATEGORY_UPDATE: async ({ args: { slug }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { name, exifTag, slug: newSlug } = await json()
     return await toCategoryApi(
       await updateCategoryLeaveWithSlug({ slug, name, exifTag, newSlug }),
     )
   },
-  CATEGORY_DELETE: async ({ args: { slug } }) => {
+  CATEGORY_DELETE: async ({ args: { slug }, context }) => {
+    if (context.role?.role !== "admin") return make403()
     try {
       await deleteCategoryLeaveWithSlug(slug)
       return true
@@ -50,10 +59,12 @@ const runHandlers = buildHandlers({
       return false
     }
   },
-  CATEGORY_CREATE: async ({ json }) => {
+  CATEGORY_CREATE: async ({ json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     return await toCategoryApi(await createCategoryLeaveWithSlug(await json()))
   },
-  CATEGORY_LIST: async ({ searchParams }) => {
+  CATEGORY_LIST: async ({ searchParams, context }) => {
+    if (!context.role) return make401()
     const p = getPaginatedParams(searchParams)
     const { content, hasMore } = await listCategories(p, {
       orphan: searchParams.has("orphan"),
@@ -64,41 +75,49 @@ const runHandlers = buildHandlers({
       items: await toCategoryApis(content),
     }
   },
-  CATEGORY_PARENT_ADD: async ({ args: { slug }, json }) => {
+  CATEGORY_PARENT_ADD: async ({ args: { slug }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { parentSlug } = await json()
     const { child } = await createCategoryLink({ parentSlug, childSlug: slug })
     return await toCategoryApi(child)
   },
-  CATEGORY_PARENT_DEL: async ({ args: { slug }, json }) => {
+  CATEGORY_PARENT_DEL: async ({ args: { slug }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { parentSlug } = await json()
     const { child } = await deleteCategoryLink({ childSlug: slug, parentSlug })
     return await toCategoryApi(child)
   },
-  CATEGORY_EXIF_REINDEX: async ({ args: { slug } }) => {
+  CATEGORY_EXIF_REINDEX: async ({ args: { slug }, context }) => {
+    if (context.role?.role !== "admin") return make403()
     await reindexCategory(slug)
     return true
   },
-  CATEGORY_BULK_PICTURE_ADD: async ({ args: { slug }, json }) => {
+  CATEGORY_BULK_PICTURE_ADD: async ({ args: { slug }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { pictureIds } = await json()
     await createPicturesCategoryLink({ slug, pictureIds })
     return true
   },
-  CATEGORY_BULK_PICTURE_DEL: async ({ args: { slug }, json }) => {
+  CATEGORY_BULK_PICTURE_DEL: async ({ args: { slug }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { pictureIds } = await json()
     await deletePicturesCategoryLink({ slug, pictureIds })
     return true
   },
-  PICTURE_UPLOAD: async ({ request }) => {
+  PICTURE_UPLOAD: async ({ request, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const formData = await request.formData()
     // @ts-expect-error
     const file: File = formData.get("file")
     const picture = await ingestPicture(file)
     return await toPictureApi(picture)
   },
-  PICTURE: async ({ args: { id } }) => {
+  PICTURE: async ({ args: { id }, context }) => {
+    if (!context.role) return make401()
     return await toPictureApi(await getPictureById(id))
   },
-  PICTURE_DELETE: async ({ args: { id } }) => {
+  PICTURE_DELETE: async ({ args: { id }, context }) => {
+    if (context.role?.role !== "admin") return make403()
     try {
       await deletePictureById(id)
       return true
@@ -108,7 +127,8 @@ const runHandlers = buildHandlers({
       return false
     }
   },
-  PICTURE_CATEGORY_ADD: async ({ args: { id }, json }) => {
+  PICTURE_CATEGORY_ADD: async ({ args: { id }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { slug } = await json()
     const picture = await getPictureById(id)
     await category_parents(db).insert({
@@ -118,7 +138,8 @@ const runHandlers = buildHandlers({
 
     return await toPictureApi(picture)
   },
-  PICTURE_CATEGORY_DEL: async ({ args: { id }, json }) => {
+  PICTURE_CATEGORY_DEL: async ({ args: { id }, json, context }) => {
+    if (context.role?.role !== "admin") return make403()
     const { slug } = await json()
     const picture = await getPictureById(id)
     await category_parents(db).delete({
@@ -128,7 +149,8 @@ const runHandlers = buildHandlers({
 
     return await toPictureApi(picture)
   },
-  PICTURE_LIST: async ({ searchParams }) => {
+  PICTURE_LIST: async ({ searchParams, context }) => {
+    if (!context.role) return make401()
     const p = getPaginatedParams(searchParams)
     const { content, hasMore } = await listPicturesPaginate(p, searchParams)
     return {
@@ -137,6 +159,8 @@ const runHandlers = buildHandlers({
     }
   },
   PICTURE_LIST_ZIP: async ({ searchParams }) => {
+    // no role check
+
     const pictures = await listPictures(searchParams)
     const archive = archiver("zip", { zlib: { level: 4 } })
     return new Response(
@@ -169,18 +193,26 @@ const runHandlers = buildHandlers({
       },
     )
   },
-  PICTURE_LIST_ZIP_PREFLIGHT: async ({ searchParams }) => {
+  PICTURE_LIST_ZIP_PREFLIGHT: async ({ searchParams, context }) => {
+    if (!context.role) return make401()
     const count = await countPictures(searchParams)
     return { pictureAmt: count, approximateSizeBytes: count * 10_000_000 }
   },
-  ADMIN_XMP_UNTRACKED: async () => {
+  ADMIN_XMP_UNTRACKED: async ({ context }) => {
+    if (context.role?.role !== "admin") return make403()
     return await getUnusedXmpTags()
   },
 })
 
-const router = createRouter()
+type RouterContext = { role: AuthRole | null }
+const router = createRouter<RouterContext>()
   .withPreMiddleware(({ request }) => {
     log(`--> ${request.url}`)
+  })
+  .withPreMiddleware(({ request, context }) => {
+    const jwt = request.headers.get("Authorization")
+    const role = jwt ? validateOrNull(jwt) : null
+    return { ...context, role }
   })
   .withPostMiddleware(({ request, response }) => {
     const s = asyncLocalStorage.getStore()
@@ -191,7 +223,7 @@ const router = createRouter()
       } DB queries`,
     )
   })
-  .run(runHandlers, null)
+  .run(runHandlers, { role: null })
 
 declare module "bun" {
   interface Env {
